backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
cloudinary==1.40.0
insightface==0.7.3
deepface==0.0.93
opencv-python-headless==4.9.0.80
numpy==1.26.4
onnxruntime==1.18.0
Pillow==10.3.0
python-dotenv==1.0.1
httpx==0.27.0
qrcode[pil]==7.4.2
moviepy==1.0.3
pydantic[email]==2.7.1
pydantic-settings==2.2.1